Allow "raw" data to be passed to Pixeldb.upsert
From a web perspective, the JSON record looks like

```
{
  "name":"a",
  "rows":3,
  "columns":3,
  "pixels":
    [
      ["#55efc4","#55efc4",null],
      ["#55efc4","#55efc4",null],
      [null,null,null]
    ]
}
```

But once Phoenix gets a hold of it, the structure
now looks more like

Based on usage between the JSON encode and the Phoenix params,
The data structure is being structured like

```
    %{
      "name" => "a",
      "rows" => "3",
      "columns" => "3",
      "pixels" => %{
        "0" => ["#55efc4", "#55efc4", ""],
        "1" => ["#55efc4", "#55efc4", ""],
        "2" => ["", "", "", ""],
      }
    }
```

So we have a few issues.

First, the `""` empty strings instead of NULL.

Second, the 2D array is now a map.

Finally, the values are all strings.

We augmented the Pixeldb to be aware of this unfortunate
difference in encoding opinion, and will handle it internally.

This will allow the web client to just "pass the buck".